backend/requirements.txt
# backend/requirements.txt

# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Validação e serialização  
pydantic==2.5.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis
redis==5.0.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Tasks assíncronas
celery==5.3.4

# Logging estruturado
structlog==23.2.0

# Utilitários
email-validator==2.1.0
python-dotenv==1.0.0